Refactor fetch-users step & tests
