+++ src/steps/__tests__/fetchUsers.test.ts
@@ -1,6 +1,6 @@
-import { createStepContext } from '../../../../test';
+import { createStepContext } from '../../../test';
 import { Recording, setupRecording } from '@jupiterone/integration-sdk-testing';
-import step from '..';
+import step from '../fetchUsers';
 
 let recording: Recording;
